int() casted to stepsReq

+++ lab5_background.py
@@ -58,7 +58,7 @@ while True:
 
     if form['angleVal'] != None:
       angle=int(form['angleVal'])
-      stepsReq=angle*(512*8)/(360) #512*8 is 1 rev in the ccw direction.
+      stepsReq=int(angle*(512*8)/(360)) #512*8 is 1 rev in the ccw direction.
       moveSteps(stepsReq,1) 
       with open('lab5_text.txt', 'w') as f:    #clear text file
         json.dump({'angleVal':0,'zerobutton':None},f)
